Service changes

+++ src/main/java/com/accenture/task/controller/CovidReportController.java
@@ -3,7 +3,6 @@ package com.accenture.task.controller;
 import com.accenture.task.constatnts.CovidConstants;
 import com.accenture.task.model.Countries;
 import com.accenture.task.service.CovidReportService;
-import com.accenture.task.service.impl.CovidReportServiceImpl;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,7 +22,7 @@ public class CovidReportController {
     @GetMapping(value = "/allCountries")
     @ApiOperation(value = "Endpoint to get Correlation Coefficient for allCountries ",
             produces = "application/json", response = Countries.class, httpMethod = "GET")
-    public float getCorrelationCoefficientForAllCountries() {
+    public double getCorrelationCoefficientForAllCountries() {
 
         return covidReportService.getCorrelationCoefficient(CovidConstants.ALL);
     }
@@ -31,7 +30,7 @@ public class CovidReportController {
     @GetMapping(value = "/continent")
     @ApiOperation(value = "Endpoint to get Correlation Coefficient by continent ",
             produces = "application/json", response = Countries.class, httpMethod = "GET")
-    public float getCorrelationCoefficientFoContinent(@RequestParam String continent) {
+    public double getCorrelationCoefficientFoContinent(@RequestParam String continent) {
 
         return covidReportService.getCorrelationCoefficient(continent);
     }
