+++ src/main/java/com/accenture/task/service/CovidReportService.java
@@ -2,5 +2,5 @@ package com.accenture.task.service;
 
 public interface CovidReportService {
 
-    public float getCorrelationCoefficient(String input);
+    public double getCorrelationCoefficient(String input);
 }
