+++ src/test/java/com/accenture/task/service/CovidReportServiceTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
-import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
@@ -15,16 +14,15 @@ public class CovidReportServiceTest {
     CovidReportService covidReportService = new CovidReportServiceImpl();
 
     @Test
-
     public void getCorrelationCoefficientForAllCountriesTest() {
-        float value = covidReportService.getCorrelationCoefficient("All");
+        double value = covidReportService.getCorrelationCoefficient("All");
         Assertions.assertNotNull(value);
 
     }
 
     @Test
     public void getCorrelationCoefficientForContinent() {
-        float value = covidReportService.getCorrelationCoefficient("Europe");
+        double value = covidReportService.getCorrelationCoefficient("Europe");
         Assertions.assertNotNull(value);
 
     }
